switch between folder mode
