+++ Wire-iOS/Sources/UserInterface/ConversationList/ListContent/ConversationListContentController/ConversationListContentController+ConversationListCellDelegate.swift
@@ -34,3 +34,9 @@ extension ConversationListContentController: ConversationListCellDelegate {
         startCallController.joinCall()
     }
 }
+
+extension ConversationListContentController {
+    func toggleSort() {
+        listViewModel.toggleSort()
+    }
+}
