+++ Wire-iOS/Sources/UserInterface/ConversationList/ListContent/ConversationListContentController/Model/ConversationListViewModel+Private.h
@@ -20,6 +20,9 @@
 
 @interface ConversationListViewModel ()
 @property (nonatomic) id<NSObject> selfUserObserver;
+
+@property (nonatomic) Boolean isFolderEnable;
+
 - (NSArray *)newConversationList;
 @end
