+++ Wire-iOS/Sources/UserInterface/ConversationList/ListContent/ConversationListContentController/Model/ConversationListViewModel+ZMConversationListObserver.swift
@@ -41,102 +41,7 @@ extension ConversationListViewModel: ZMConversationListObserver {
             let sectionIndex = SectionIndex.contactRequests
             updateSection(sectionIndex)
 
-            delegate?.listViewModel(self, didUpdateSectionForReload: sectionIndex.uIntValue)
+            delegate?.listViewModel(self, didUpdateSectionForReload: sectionIndex.sectionNumber(isFolderEnable: isFolderEnable))
         }
     }
-}
-
-
-extension ConversationListViewModel {
-
-    @objc
-    func updateAllSections() {
-        updateSection(sectionIndexs: SectionIndex.allCases)
-    }
-
-    func updateSection(_ sectionIndex: SectionIndex, withItems items: [Any]? = nil) {
-        updateSection(sectionIndexs: [sectionIndex], withItems: items)
-    }
-
-    /// This updates a specific section in the model, by copying the contents locally.
-    /// Passing in a value of SectionIndexAll updates all sections. The reason why we need to keep
-    /// local copies of the lists is that we get separate notifications for each list,
-    /// which means that an update to one can render the collection view out of sync with the datasource.
-    ///
-    /// - Parameter sectionIndex: the section to update
-    func updateSection(sectionIndexs: [SectionIndex], withItems items: [Any]? = nil) {
-        if sectionIndexs == SectionIndex.allCases && items != nil {
-            assert(true, "Update for all sections with proposed items is not allowed.")
-        }
-
-        for sectionIndex in sectionIndexs {
-            switch sectionIndex {
-            case .contactRequests:
-                if let userSession = ZMUserSession.shared(), ZMConversationList.pendingConnectionConversations(inUserSession: userSession).count > 0 {
-                    inbox = items ?? [self.contactRequestsItem]
-                } else {
-                    inbox = []
-                }
-            case .conversations:
-                // Make a new copy of the conversation list
-                conversations = items ?? newConversationList()
-            case .contactsConversations:
-                oneOnOneConversations = items ?? newOneOnOneConversationList()
-            }
-        }
-
-
-        ///TODO: use a dictionary instead of fix size array
-        // Re-create the aggregate array
-        let sections: [Any] = [inbox ?? [],
-                               oneOnOneConversations ?? [],
-                               conversations ?? []]
-
-        aggregatedItems = AggregateArray(sections: sections)
-    }
-
-    @discardableResult
-    func updateForConvoType(sectionIndex: SectionIndex) -> Bool {
-        if let oldConversationList = aggregatedItems.section(at: sectionIndex.uIntValue) as? Array<AnyHashable>,
-            let newConversationList = newConversationList() as? Array<AnyHashable>,
-            oldConversationList != newConversationList {
-
-            let startState = ZMOrderedSetState(orderedSet: NSOrderedSet(array: oldConversationList))
-            let endState = ZMOrderedSetState(orderedSet: NSOrderedSet(array: newConversationList))
-            let updatedState = ZMOrderedSetState(orderedSet: [])
-
-            guard let changedIndexes = ZMChangedIndexes(start: startState, end: endState, updatedState: updatedState, moveType: ZMSetChangeMoveType.uiCollectionView) else { return true}
-
-            if changedIndexes.requiresReload == true {
-                reload()
-            } else {
-                // We need to capture the state of `newConversationList` to make sure that we are updating the value
-                // of the list to the exact new state.
-                // It is important to keep the data source of the collection view consistent, since
-                // any inconsistency in the delta update would make it throw an exception.
-                let modelUpdates = {
-                    self.updateSection(sectionIndex, withItems: newConversationList)
-                }
-                delegate?.listViewModel(self, didUpdateSection: sectionIndex.uIntValue, using: modelUpdates, with: changedIndexes)
-            }
-
-            return true
-        }
-
-        return false
-    }
-
-    func updateConversationListAnimated() {
-        if numberOfItems(inSection: SectionIndex.conversations.uIntValue) == 0 &&
-           numberOfItems(inSection: SectionIndex.contactsConversations.uIntValue) == 0 {
-            reload()
-        } else if !updateForConvoType(sectionIndex: .conversations) {
-            updateForConvoType(sectionIndex: .contactsConversations)
-        }
-    }
-
-    func newOneOnOneConversationList() -> [ZMConversation]? {
-        return ZMUserSession.shared()?.oneOnOneConversations.map { $0 }
-    }
-
 }
